update proftp
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@17876 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/proftpd/include/netio.h
@@ -2,7 +2,7 @@
  * ProFTPD - FTP server daemon
  * Copyright (c) 1997, 1998 Public Flood Software
  * Copyright (c) 1999, 2000 MacGyver aka Habeeb J. Dihu <macgyver@tos.net>
- * Copyright (c) 2001-2008 The ProFTPD Project
+ * Copyright (c) 2001-2011 The ProFTPD Project
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,7 +16,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307, USA.
+ * Foundation, Inc., 51 Franklin Street, Suite 500, Boston, MA 02110-1335, USA.
  *
  * As a special exemption, Public Flood Software/MacGyver aka Habeeb J. Dihu
  * and other respective copyright holders give permission to link this program
@@ -25,7 +25,7 @@
  */
 
 /* Network IO stream layer
- * $Id: netio.h,v 1.10 2008/01/15 17:56:33 castaglia Exp $
+ * $Id: netio.h,v 1.12 2011/05/23 20:35:35 castaglia Exp $
  */
 
 #ifndef PR_NETIO_H
@@ -82,7 +82,7 @@ typedef struct {
   char *current;
 
   /* Number of bytes left in the buffer. */
-  int remaining;
+  size_t remaining;
 
 } pr_buffer_t;
